EC: Move prime numbers to primes.py
